render pass

+++ src/app/renderer_new/mod.rs
@@ -2,4 +2,4 @@ mod free_list;
 mod renderer_new;
 mod vertex_arena;
 mod vm;
-static CHUNK_SIZE: u64 = 1024 * 4;
+static CHUNK_SIZE: u32 = 1024 * 4;
